backend/requirements.txt
# Web frameworks
Flask
Flask-Cors
fastapi
uvicorn

# Flask dependencies
python-dotenv
Jinja2
MarkupSafe
itsdangerous
Werkzeug
click

# Auth
PyJWT

# AI/LLM
openai
langchain
langchain-core
langchain-community
langchain-aws

# AWS
boto3

# Data
pymongo
python-multipart
requests
beautifulsoup4